requirements.txt
torch
torchvision
transformers
opencv-python
faiss-cpu
Pillow
numpy
yt-dlp
streamlit
ftfy
regex
tqdm
